RM-55013 Refactor: solve circular dependency and use proper inteface/implementation for EbicsUserService
