+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/module/BankPaymentModule.java
@@ -54,6 +54,8 @@ import com.axelor.apps.bankpayment.ebics.service.EbicsBankService;
 import com.axelor.apps.bankpayment.ebics.service.EbicsBankServiceImpl;
 import com.axelor.apps.bankpayment.ebics.service.EbicsPartnerService;
 import com.axelor.apps.bankpayment.ebics.service.EbicsPartnerServiceImpl;
+import com.axelor.apps.bankpayment.ebics.service.EbicsUserService;
+import com.axelor.apps.bankpayment.ebics.service.EbicsUserServiceImpl;
 import com.axelor.apps.bankpayment.service.AccountingReportPrintServiceBankPaymentImpl;
 import com.axelor.apps.bankpayment.service.PaymentScheduleLineBankPaymentService;
 import com.axelor.apps.bankpayment.service.PaymentScheduleLineBankPaymentServiceImpl;
@@ -148,6 +150,8 @@ public class BankPaymentModule extends AxelorModule {
     bind(MoveRemoveService.class).to(MoveRemoveServiceBankPaymentImpl.class);
 
     bind(EbicsUserRepository.class).to(EbicsUserManagementRepository.class);
+
+    bind(EbicsUserService.class).to(EbicsUserServiceImpl.class);
 
     bind(AccountingReportPrintServiceImpl.class)
         .to(AccountingReportPrintServiceBankPaymentImpl.class);
